[QOL-7762] refactor - split resource type validation into a separate module
- This should make it easier to review, maintain, and possibly share with others
